FF-314 #comment Facets file/module; expset style
